+++ dkz/src/math/my_rounds.c
@@ -5,7 +5,7 @@
 ** my_rounds
 */
 
-int my_ceil(double x)
+double my_ceil(double x)
 {
     int i = (long long)x;
 
@@ -14,7 +14,7 @@ int my_ceil(double x)
     return i;
 }
 
-int my_floor(double x)
+double my_floor(double x)
 {
     int i = (long long)x;
 
@@ -23,7 +23,7 @@ int my_floor(double x)
     return i;
 }
 
-int my_round(double x)
+double my_round(double x)
 {
     if (x < 0)
         return my_ceil(x - 0.5);
@@ -31,7 +31,7 @@ int my_round(double x)
         return my_floor(x + 0.5);
 }
 
-int my_trunc(double x)
+double my_trunc(double x)
 {
     return (long long)x;
 }
